sol2

+++ Fracture/src/Component/CameraControllerComponent.cpp
@@ -72,8 +72,8 @@ void Fracture::CameraControllerComponent::Move(Camera_Movement td, float dt)
 bool firstMouse = true;
 void Fracture::CameraControllerComponent::InputMouse(float xpos, float ypos, float dt, bool constrainPitch)
 {
-    static double lastX = xpos;
-    static double lastY = ypos;
+    static float lastX = xpos;
+    static float lastY = ypos;
     
     float xoffset = xpos - lastX;
     float yoffset = lastY - ypos; // reversed since y-coordinates go from bottom to top
